requirements.txt
alembic==1.7.1
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
automaton==2.4.0
autopage==0.4.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.18.23
botocore==1.21.23
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cliff==3.9.0
cmd2==2.1.2
colorama==0.4.4
cryptography==3.4.8
dataclasses==0.8
debtcollector==2.3.0
decorator==4.4.2
dnspython==2.1.0
docker==5.0.2
dogpile.cache==1.1.4
eventlet==0.32.0
extras==1.0.0
fastapi==0.68.1
fasteners==0.16.3
fixtures==3.0.0
Flask==2.0.1
futurist==2.4.0
google-auth==2.0.2
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.2
importlib-metadata==4.6.4
importlib-resources==5.2.2
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
keystoneauth1==4.3.1
keystonemiddleware==9.3.0
kombu==5.1.0
kubernetes==18.20.0
logutils==0.3.5
Mako==1.1.5
mangum==0.12.2
marathon==0.13.0
MarkupSafe==2.0.1
msgpack==1.0.2
munch==2.5.0
netaddr==0.8.0
netifaces==0.11.0
networkx==2.5.1
oauthlib==3.1.1
openstacksdk==0.59.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==2.4.2
oslo.cache==2.8.1
oslo.concurrency==4.4.1
oslo.config==8.7.1
oslo.context==3.3.1
oslo.db==11.0.0
oslo.i18n==5.1.0
oslo.log==4.6.0
oslo.messaging==12.9.1
oslo.metrics==0.3.0
oslo.middleware==4.4.0
oslo.policy==3.8.2
oslo.reports==2.3.0
oslo.serialization==4.2.0
oslo.service==2.6.0
oslo.upgradecheck==1.4.0
oslo.utils==4.10.0
oslo.versionedobjects==2.5.0
packaging==21.0
Paste==3.5.0
PasteDeploy==2.1.1
pbr==5.6.0
pecan==1.4.0
prettytable==2.2.0
prometheus-client==0.11.0
psutil==5.8.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycadf==3.1.1
pycparser==2.20
pydantic==1.8.2
pydot==1.4.2
pyinotify==0.9.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-barbicanclient==5.1.0
python-cinderclient==8.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-glanceclient==3.5.0
python-heatclient==2.4.0
python-keystoneclient==4.2.0
python-neutronclient==7.6.0
python-novaclient==17.6.0
python-octaviaclient==2.4.0
python-openstackclient==5.6.0
python-swiftclient==3.12.0
pytz==2021.1
PyYAML==5.4.1
repoze.lru==0.7
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requestsexceptions==1.4.0
rfc3986==1.5.0
Routes==2.5.1
rsa==4.7.2
s3transfer==0.5.0
simplegeneric==0.8.1
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.23
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
starlette==0.14.2
statsd==3.3.0
stevedore==3.4.0
taskflow==4.6.2
Tempita==0.5.2
tenacity==8.0.1
testresources==2.0.1
testscenarios==0.5.0
testtools==2.5.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.14.0
vine==5.0.0
waitress==2.0.0
warlock==1.3.3
watchgod==0.7
wcwidth==0.2.5
WebOb==1.8.7
websocket-client==1.2.1
websockets==9.1
WebTest==3.0.0
Werkzeug==2.0.1
wrapt==1.12.1
WSME==0.11.0
yappi==1.3.2
zipp==3.5.0
